Remove favicon hack
The hack caused the page to be fetched twice! That's not good.
Any other "fixes" that I've found will either: respond with a
transparent GIF or PNG which results in a gap in the tab, respond with
no data which will cause the browser to display their fallback icons.
I don't really like either of those options so I chose to have the
browsers look for /favicon.ico and receive a 404 instead of fetching
page's HTML twice. eh.